Apply the TMPS model to a set of data for a typewise statistical estimate of return periods as well as a combined TMPS return period. See reference Fischer (2018)

Apply the TMPS model to a set of data for a typewise statistical estimate of return periods as well as a combined TMPS return period. See reference Fischer (2018)

TMPS_model(
  Flood_events,
  Daily_discharge,
  return_TMPS,
  Threshold_Q,
  p_input = NULL,
  q_input = NULL
)

Arguments

Flood_events

data.frame: Floods events with the columns "Peak_date" (Date format), "HQ" (numeric) and "Type" (factor)

Daily_discharge

data.frame: Daily discharge (continuous) with columns "Date (Date format) and "Discharge" (numeric)

return_TMPS

character vector: What should be returned. Any combination of "TMPS" and the occuring floodtypes

Threshold_Q

numeric: Peak of Threshold, defaults to 3

p_input

The points (as years of return period) where the quantiles should be calculated

q_input

Probabilities to calculate the quantiles for

Author

Svenja Fischer

Philipp Bühler