Apply the TMPS model for a typewise statistical estimate of return periods as well as a combined TMPS return period. See also reference Fischer (2018).

Apply the TMPS model for a typewise statistical estimate of return periods as well as a combined TMPS return period. See also reference Fischer (2018).

qTMPS(
  p = c(2, 5, 10, 20, 25, 50, 100, 200, 500, 1000),
  Flood_events,
  Daily_discharge,
  return_TMPS = c("TMPS", "R1", "R2", "R3", "S1", "S2"),
  Threshold_Q = 3,
  p_as_annuality = TRUE
)

Arguments

p

The points (return periods) where the quantiles should be calculated Defaults to: 2, 5, 10, 20, 25, 50, 100, 200, 500, 1000

Flood_events

data.frame: Floods events with the columns "Peak_date" (Date format), "HQ" (numeric) and "Type" (factor)

Daily_discharge

data.frame: Daily discharge (continuous) with columns "Date (Date format) and "Discharge" (numeric)

return_TMPS

character vector: What should be returned. Any combination of "TMPS" and the occuring floodtypes

Threshold_Q

numeric: Peak of Threshold, defaults to 3

p_as_annuality

logical: Should all probabilities be treated as an annuality [in years] instead of dimensionless values between 0 and 1

References

Fischer, S. (2018). A seasonal mixed-POT model to estimate high flood quantiles from different event types and seasons. Journal of Applied Statistics, 45(15), 2831–2847. https://doi.org/10.1080/02664763.2018.1441385

Author

Svenja Fischer

Philipp Bühler